requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dataclasses-json==0.6.7
distro==1.9.0
et_xmlfile==2.0.0
faiss-cpu==1.11.0
fastapi==0.115.12
frozenlist==1.6.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.3.42
lxml==5.4.0
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.77.0
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
propcache==0.3.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
XlsxWriter==3.2.3
yarl==1.20.0
zstandard==0.23.0
